game-screenshot-tool/requirements.txt
keyboard>=0.13.5
Pillow>=10.0.0
pystray>=0.19.0
win10toast>=0.9
pyinstaller>=6.0
